requirements-api.txt
attr==0.3.2
Flask==3.0.3
Flask_Cors==4.0.1
Flask_HTTPAuth==4.8.0
Flask_Injector==0.15.0
Flask_RESTful==0.3.10
injector==0.21.0
numpy==1.26.4
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
PyHamcrest==2.1.0
pymongo==4.7.2
python_dateutil==2.8.2  # Corrected to python-dateutil, but kept your version for compatibility
Requests==2.32.2  # Updated to maintain the latest mentioned version
torch==1.13.1
Werkzeug==3.0.3  # Kept higher version from your original list
attrs==21.4.0
toolz==0.11.2
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.3
bson==0.5.10
gunicorn==20.1.0
jsonschema==3.2.0
cattrs==1.8.0
workalendar==16.3.0
pyhumps==3.7.3
humps==0.2.2
decamelize==0.1.2
beautifulsoup4==4.11.1
annotated-types==0.5.0
pydantic==2.0.2
pydantic-core==2.1.2


libclang==18.1.1